requirements.txt
streamlit
Authlib
supabase
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

pandas 
streamlit-aggrid
